pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "scholar-fetch"
version = "0.3.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "openai>=1.30",
    "openreview-py>=1.59.1",
    "pydantic>=2.7",
    "typer>=0.12",
    "rapidfuzz>=3.9",
    "tenacity>=9.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.scripts]
scholar = "scholar_fetch.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["scholar_fetch*"]

[tool.setuptools.package-data]
scholar_fetch = ["data/*.yaml"]
